fix(runtime): move AskError to internal::types to unblock WASM build
Root cause: actor.rs imported AskError via `use crate::hew_node::AskError`,
but hew_node is gated #[cfg(not(target_arch = "wasm32"))] in lib.rs,
causing an unresolved import on wasm32-wasip1 --no-default-features.

Fix: move the AskError enum definition from hew_node.rs into
hew-runtime/src/internal/types.rs, which is unconditionally compiled on
all targets.  hew_node.rs now re-exports it with `pub use` for backward
compatibility.  actor.rs imports directly from internal::types so it
compiles on both native and WASM.

Validated:
  cargo build -p hew-runtime --target wasm32-wasip1 --no-default-features --release  -> Finished (0 errors)
  cargo test  -p hew-runtime --lib -- --test-threads=1                                -> 1064 passed, 0 failed
